fix double
git-svn-id: http://svn.terracotta.org/svn/ehcache@3246 b9324663-ca0f-0410-8574-be9b3887307d

+++ trunk/core/src/main/java/net/sf/ehcache/search/aggregator/Average.java
@@ -25,7 +25,7 @@ import net.sf.ehcache.search.Attribute;
  */
 public class Average implements AggregatorInstance<Double> {
 
-    private long count;
+    private double count;
     private double sum;
     private final Attribute<?> attribute;
